Redo the biggest asset to reduce its size.
And by biggest I mean the file size is larger than everything else.

This basically increase the compression and trim useless parts.

And yes, i had to edit the entire history to remove the old version,
because uploading it would be 100% useless and use precious disk space
from my free github plan, as well as yours.