Reimplement a media loader in javascript, because of CORS issues.
When running in nwjs in a ccmod pack, many media fails to load because
the service worker isn't allowed to do everything it wants when using
medias.  The common workaround is to download everything as a blob or
bufferarray and then stuff it into some API that is happy with it.