Fix ccmod.json compatibility with ccloader3
ccmod.json was initially implemented by ccloader3.  It used to have no
specification (as usual) so the only point of reference was other
ccloader3 mods like crosscode-ru and the ccloader3 source code.

ccloader2 eventually supported ccmod.json.  Lately, a new ccmod.json
specification is being written¹ which adds a new useful fields: "tags".

This specification allows the "authors" field to be a string, but
ccloader3 fails if this field is not an array.

[1]: https://github.com/CCDirectLink/CCModDB/blob/39e0db315c72869c88ebb2767fcc2cefd1123ebc/CCMOD-STANDARD.md